+++ CHANGES.rst
@@ -27,6 +27,8 @@ Imviz
 
 - Aperture photometry (previously "Imviz Simple Aperture Photometry") now supports batch mode. [#2465]
 
+- Aperture photometry sum is now presented in scientific notation consistently. [#2530]
+
 - Expose sky regions in get_subsets. If 'include_sky_region' is True, a sky Region will be returned (in addition to a pixel Region) for spatial subsets with parent data that was a WCS. [#2496]
 
 Mosviz